+++ src/language_name.rs
@@ -20,6 +20,7 @@ impl LanguageName {
     pub const CYTHON: Self = Self("Cython");
     pub const TEX: Self = Self("TeX");
     pub const YAML: Self = Self("YAML");
+    pub const IPYNB: Self = Self("Jupyter Notebook");
 
     /// Convert language names from the strings we used in the v1 and v2 config format
     pub fn from_legacy(s: &str) -> Result<Self, UnknownLanguageError> {